Applied Command Design Pattern
In the aardvark Class line, no 835 to 846 has Divergent Change. It indicates that there is an irrelevant method in the aardvark class. That is called as drawDominoesGuesses(). When considering that method, it is most appropriate for pictureframe class. I moved it into the pictureframe class. In PictureFrame class line no:- 120 to 130.